+++ workspace/plugins/inbox/src/vite-env.d.ts
@@ -1,2 +1,4 @@
 /// <reference types="svelte" />
 /// <reference types="vite/client" />
+/// <reference types="vite-plugin-pwa/vue" />
+/// <reference types="@workspace/types/global" />