+++ workspace/plugins/inbox/vite.share.config.ts
@@ -1,41 +1,35 @@
-/**
- * Copyright (C) 2023 Zuoqiu Yingyi
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- * 
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+// Copyright (C) 2023 Zuoqiu Yingyi
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+// 
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { defineConfig } from "vite";
-import { resolve } from "node:path";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
 import { less } from "svelte-preprocess-less";
+import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vitejs.dev/config/
 export default defineConfig({
     base: `./`,
     plugins: [
+        // REF: https://www.npmjs.com/package/vite-tsconfig-paths
+        tsconfigPaths(),
         svelte({
             preprocess: {
                 style: less(),
             },
         }),
     ],
-    resolve: {
-        alias: {
-            "~": resolve(__dirname, "./"),
-            "@": resolve(__dirname, "./src"),
-        }
-    },
     build: {
         minify: true,
         // sourcemap: "inline",
@@ -47,21 +41,21 @@ export default defineConfig({
                 /^@electron\/.*$/,
             ],
             output: {
-                entryFileNames: entryInfo => {
+                entryFileNames: (entryInfo) => {
                     // console.log(entryInfo);
                     switch (entryInfo.name) {
                         default:
                             return "assets/[name]-[hash].js";
                     }
                 },
-                chunkFileNames: chunkInfo => {
+                chunkFileNames: (chunkInfo) => {
                     // console.log(chunkInfo);
                     switch (chunkInfo.name) {
                         default:
                             return "chunks/[name]-[hash].js";
                     }
                 },
-                assetFileNames: assetInfo => {
+                assetFileNames: (assetInfo) => {
                     // console.log(chunkInfo);
                     switch (assetInfo.name) {
                         default: